Add fatal error information to root span on PHP 7
Add ddtrace_exception_to_meta API
Add ddtrace_assert_log_debug macro and fix logging macros.

Note I found a new PHP 8.0 incompatibility:
add_assoc_zval_ex is now a void function. It's just a proxy to
another function, so let's just use that.
Note that on 7.3+ this will always succeed, but I'm not _certain_
that on <7.2 that it won't fail, which is the reason for checking
the retval.

Split ddtrace_error_cb in to version files