Fix Symfony 4.0 tests
This updates them to the latest changes made to the integration,
but also changes how this version is installed. We now commit the
composer.lock and do `composer install` instead of an update to
avoid broken dependencies in this unmaintained version.

Side bonus: the composer install is much faster than update.